[ALPS05232861] p2p: Fix type mismatch
Fix type mismatch

MTK-Commit-Id: e23094f229d3b46bf92acf8f658f79f4af996a3f

CR-Id: ALPS05232861
Feature: [Module] Wi-Fi Driver

+++ os/linux/gl_p2p.c
@@ -1226,7 +1226,7 @@ int glSetupP2P(struct GLUE_INFO *prGlueInfo, struct wireless_dev *prP2pWdev,
 		"check prDevHandler = %p, aprRoleHandler = %p\n",
 		prP2PInfo->prDevHandler, prP2PInfo->aprRoleHandler);
 
-	prNetDevPriv->ucBssIdx = p2pRoleFsmInit(prAdapter, u4Idx);
+	prNetDevPriv->ucBssIdx = p2pRoleFsmInit(prAdapter, (uint8_t) u4Idx);
 	init_completion(&prP2PInfo->rStopApComp);
 	/* Currently wpasupplicant can't support create interface. */
 	/* so initial the corresponding data structure here. */
